# docs/swagger.yaml
definitions:
  model.HTTPError400:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  model.HTTPError500:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Internal Server Error
        type: string
    type: object
info:
  contact:
    email: ahmad.hassan@gmail.com
    name: Dr Ahmad Hassan
    url: https://www.linkedin.com/in/ahmadhassan
  description: This is an api for notification service.
  termsOfService: http://swagger.io/terms/
  title: Notification API
  version: "1.0"
paths:
  /notification/activate:
    post:
      consumes:
      - application/json
      description: todo
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError500'
      summary: Activate the notification
      tags:
      - Notification
  /notification/subscribe:
    post:
      consumes:
      - application/json
      description: todo
      parameters:
      - description: The notification url
        in: query
        name: url
        required: true
        type: string
      - description: The id of the customer
        in: query
        name: customerId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError500'
      summary: Test the notification
      tags:
      - Notification
  /notification/test:
    post:
      consumes:
      - application/json
      description: todo
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPError400'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError500'
      summary: Test the notification
      tags:
      - Notification
swagger: "2.0"